+++ lib/api.dart
@@ -8,7 +8,7 @@ import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 import 'package:freezed_annotation/freezed_annotation.dart' hide protected;
 part 'api.freezed.dart';
 
-// These functions are ignored because they are not marked as `pub`: `get_credentials`, `git_add_all_impl`, `git_add_remote_impl`, `git_checkout_impl`, `git_commit_impl`, `git_diff_impl`, `git_fetch_impl`, `git_init_impl`, `git_merge_abort_impl`, `git_merge_prefer_remote_impl`, `git_pull_impl`, `git_push_impl`, `git_stash_impl`, `git_status_impl`
+// These functions are ignored because they are not marked as `pub`: `get_credentials`, `git_add_all_impl`, `git_add_remote_impl`, `git_checkout_impl`, `git_commit_impl`, `git_diff_impl`, `git_fetch_impl`, `git_init_impl`, `git_merge_abort_impl`, `git_merge_prefer_remote_impl`, `git_pull_impl`, `git_push_impl`, `git_stash_impl`, `git_status_impl`, `validate_certificate`
 // These function are ignored because they are on traits that is not defined in current crate (put an empty `#[frb]` on it to unignore): `fmt`, `fmt`, `from`, `from`
 
 Future<int> add({required int left, required int right}) =>
@@ -105,6 +105,9 @@ Future<String> gitStash({required String path}) =>
 
 Future<String> gitDiff({required String path}) =>
     RustLib.instance.api.crateApiGitDiff(path: path);
+
+Future<void> setSslCaCerts({required List<String> pemCerts}) =>
+    RustLib.instance.api.crateApiSetSslCaCerts(pemCerts: pemCerts);
 
 @freezed
 sealed class GitError with _$GitError implements FrbException {